requirements.txt
Flask
python-dotenv
google-genai
gTTS
requests
gradio_client
